+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The **third number** is the patch version (bug fixes)
 - Command `commands`: list all available commands.
 - Option: `user list --sort [id|username|name]`.
 - `UserResp` compact table format (used by `user list`).
+- `--paging`, `--pager` options and corresponding config options for displaying results in a pager.
 
 ### Changed